rename old component
