+++ ui/src/pages/namespace/Settings/Secrets/CreateEdit_obsolete.tsx
